+++ example/transaction/processMoneyTransfer.js
@@ -7,32 +7,14 @@ async function main() {
     await mongoose.connect(process.env.MONGOLAB_URI);
 
     const actionRepo = new domain.repository.Action(mongoose.connection);
+    const projectRepo = new domain.repository.Project(mongoose.connection);
     const transactionRepo = new domain.repository.Transaction(mongoose.connection);
 
-    const accountService = new domain.pecorinoapi.service.Account({
-        endpoint: process.env.PECORINO_ENDPOINT,
-        auth: new domain.pecorinoapi.auth.ClientCredentials({
-            domain: process.env.PECORINO_AUTHORIZE_SERVER_DOMAIN,
-            clientId: process.env.PECORINO_CLIENT_ID,
-            clientSecret: process.env.PECORINO_CLIENT_SECRET,
-            scopes: [],
-            state: ''
-        })
-    })
-
-    const transferTransactionService = new domain.pecorinoapi.service.transaction.Transfer({
-        endpoint: process.env.PECORINO_ENDPOINT,
-        auth: new domain.pecorinoapi.auth.ClientCredentials({
-            domain: process.env.PECORINO_AUTHORIZE_SERVER_DOMAIN,
-            clientId: process.env.PECORINO_CLIENT_ID,
-            clientSecret: process.env.PECORINO_CLIENT_SECRET,
-            scopes: [],
-            state: ''
-        })
-    });
-
     const transaction = await domain.service.transaction.moneyTransfer.start({
-        expires: moment().add(5, 'minutes').toDate(),
+        project: { typeOf: domain.factory.chevre.organizationType.Project, id: 'cinerino' },
+        expires: moment()
+            .add(1, 'minutes')
+            .toDate(),
         agent: {
             typeOf: domain.factory.personType.Person,
             id: 'personId',
@@ -45,35 +27,30 @@ async function main() {
         },
         object: {
             clientUser: {},
-            amount: 10,
-            toLocation: { accountType: domain.factory.accountType.Point, accountNumber: '50031310908' }
-        }
+            amount: { value: 10, currency: 'JPY' },
+            fromLocation: { typeOf: 'PrepaidPaymentCard', identifier: '139485855034733' },
+            toLocation: { typeOf: 'PrepaidPaymentCard', identifier: '430458588634915' },
+            description: 'サンプル転送取引'
+        },
+        seller: { typeOf: 'Corporation', id: '59d20831e53ebc2b4e774466' },
     })({
-        accountService: accountService,
+        action: actionRepo,
+        project: projectRepo,
         transaction: transactionRepo
     });
     console.log('transaction started', transaction);
 
-    const authorizeAccountPaymentAction = await domain.service.payment.account.startTransaction({
-        object: {
-            typeOf: domain.factory.paymentMethodType.Account,
-            amount: transaction.object.amount,
-            fromAccount: { accountType: domain.factory.accountType.Point, accountNumber: '60101119300' },
-            toAccount: { accountType: domain.factory.accountType.Point, accountNumber: '50031310908' },
-            notes: 'test from samples',
-            currency: domain.factory.accountType.Point
-        },
-        agent: transaction.agent,
-        purpose: transaction
+    await domain.service.transaction.updateAgent({
+        typeOf: transaction.typeOf,
+        id: transaction.id,
+        agent: { id: transaction.agent.id, name: `セイ　メイ`, email: 'hello@motionpicture.jp', telephone: '+819012345678' }
     })({
-        action: actionRepo,
-        transaction: transactionRepo,
-        transferTransactionService: transferTransactionService
+        transaction: transactionRepo
     });
-    console.log('account authorized', authorizeAccountPaymentAction);
+    console.log('profile set');
 
     try {
-        await domain.service.transaction.moneyTransfer.confirm(transaction)({
+        await domain.service.transaction.moneyTransfer.confirm({ id: transaction.id })({
             action: actionRepo,
             transaction: transactionRepo
         });