+++ src/service/payment/chevre.ts
@@ -200,19 +200,44 @@ export function voidPayment(params: factory.task.IData<factory.taskName.VoidPaym
             );
         }
 
+        const transactionService = new chevre.service.Transaction({
+            endpoint: credentials.chevre.endpoint,
+            auth: chevreAuthClient
+        });
         const payService = new chevre.service.transaction.Pay({
             endpoint: credentials.chevre.endpoint,
             auth: chevreAuthClient
         });
 
+        const errors: any[] = [];
         for (const action of authorizeActions) {
             // 直列にゆっくり処理する場合↓
             // tslint:disable-next-line:no-magic-numbers
             // await new Promise((resolve) => setTimeout(() => { resolve(); }, 1000));
 
-            await payService.cancel({ transactionNumber: action.object.paymentMethodId });
+            // 失敗するケースがあっても、残りが少なくとも処理されるようにエラーハンドリング
+            try {
+                // 取引が存在すれば中止
+                const transactionNumber = action.object.paymentMethodId;
+                if (typeof transactionNumber === 'string' && transactionNumber.length > 0) {
+                    const { data } = await transactionService.search({
+                        limit: 1,
+                        project: { ids: [action.project.id] },
+                        typeOf: chevre.factory.transactionType.Pay,
+                        transactionNumber: { $eq: transactionNumber }
+                    });
+                    if (data.length > 0) {
+                        await payService.cancel({ transactionNumber });
+                    }
+                }
 
-            await repos.action.cancel({ typeOf: action.typeOf, id: action.id });
+                await repos.action.cancel({ typeOf: action.typeOf, id: action.id });
+            } catch (error) {
+                errors.push(error);
+            }
+        }
+        if (errors.length > 0) {
+            throw errors[0];
         }
     };
 }