+++ src/service/transaction/moneyTransfer.ts
@@ -14,6 +14,7 @@ import { MongoRepository as TaskRepo } from '../../repo/task';
 import { MongoRepository as TransactionRepo } from '../../repo/transaction';
 
 import { createPotentialActions } from './moneyTransfer/potentialActions';
+import { IPassportValidator as IWaiterPassportValidator, validateWaiterPassport } from './validation';
 
 import { handleChevreError } from '../../errorHandler';
 
@@ -42,13 +43,16 @@ export type IConfirmOperation<T> = (repos: {
     transaction: TransactionRepo;
 }) => Promise<T>;
 
+export type IPassportValidator = IWaiterPassportValidator;
+export type IStartParams = factory.transaction.moneyTransfer.IStartParamsWithoutDetail & {
+    passportValidator?: IPassportValidator;
+};
+
 /**
  * 取引開始
  * Chevre通貨転送サービスを利用して転送取引を開始する
  */
-export function start(
-    params: factory.transaction.moneyTransfer.IStartParamsWithoutDetail
-): IStartOperation<factory.transaction.moneyTransfer.ITransaction> {
+export function start(params: IStartParams): IStartOperation<factory.transaction.moneyTransfer.ITransaction> {
     return async (repos: {
         action: ActionRepo;
         project: ProjectRepo;
@@ -60,10 +64,12 @@ export function start(
         });
         const seller = await sellerService.findById({ id: params.seller.id });
 
+        const passport = await validateWaiterPassport(params);
+
         // 金額をfix
         const amount = params.object.amount;
-        if (typeof amount !== 'number') {
-            throw new factory.errors.ArgumentNull('amount');
+        if (typeof amount.value !== 'number') {
+            throw new factory.errors.ArgumentNull('amount.value');
         }
 
         // fromとtoをfix
@@ -91,7 +97,10 @@ export function start(
                 fromLocation: fromLocation,
                 toLocation: toLocation,
                 authorizeActions: [],
-                ...(typeof params.object.description === 'string') ? { description: params.object.description } : {}
+                ...(passport !== undefined) ? { passport } : undefined,
+                ...(typeof params.object.description === 'string') ? { description: params.object.description } : undefined,
+                ...(typeof params.object.pendingTransaction?.identifier === 'string')
+                    ? { pendingTransaction: { identifier: params.object.pendingTransaction.identifier } } : undefined
             },
             expires: params.expires
         };
@@ -125,47 +134,50 @@ function authorizePaymentCard(params: {
         transaction: TransactionRepo;
     }) => {
         const transaction = params.transaction;
-        // const amount = transaction.object.amount;
-
-        const fromLocation = <factory.action.transfer.moneyTransfer.IPaymentCard>transaction.object.fromLocation;
+        const fromLocation = transaction.object.fromLocation;
 
         if (typeof transaction.object.toLocation.typeOf === 'string') {
-            const toLocation = <factory.action.transfer.moneyTransfer.IPaymentCard>transaction.object.toLocation;
+            const toLocation = transaction.object.toLocation;
 
             // 転送取引
             await processAuthorizePaymentCard({
                 project: { typeOf: transaction.project.typeOf, id: transaction.project.id },
                 agent: { id: transaction.agent.id },
                 object: {
-                    amount: transaction.object.amount,
-                    typeOf: factory.action.authorize.paymentMethod.any.ResultType.Payment,
-                    // typeOf: fromLocation.typeOf,
-                    paymentMethod: fromLocation.typeOf,
+                    project: { typeOf: transaction.project.typeOf, id: transaction.project.id },
+                    typeOf: factory.chevre.offerType.AggregateOffer,
+                    itemOffered: {
+                        typeOf: 'MonetaryAmount',
+                        value: transaction.object.amount.value,
+                        currency: transaction.object.amount.currency
+                    },
                     fromLocation: fromLocation,
                     toLocation: toLocation,
-                    ...{
-                        description: transaction.object.description
-                    }
+                    seller: transaction.seller,
+                    price: 0,
+                    priceCurrency: factory.priceCurrency.JPY,
+                    description: transaction.object.description
                 },
                 purpose: { typeOf: transaction.typeOf, id: transaction.id }
             })(repos);
         } else {
+            throw new factory.errors.NotImplemented('Withdraw transaction not implemented');
             // 出金取引
-            await processAuthorizePaymentCard({
-                project: { typeOf: transaction.project.typeOf, id: transaction.project.id },
-                agent: { id: transaction.agent.id },
-                object: {
-                    amount: transaction.object.amount,
-                    typeOf: factory.action.authorize.paymentMethod.any.ResultType.Payment,
-                    // typeOf: fromLocation.typeOf,
-                    paymentMethod: fromLocation.typeOf,
-                    fromLocation: fromLocation,
-                    ...{
-                        description: transaction.object.description
-                    }
-                },
-                purpose: { typeOf: transaction.typeOf, id: transaction.id }
-            })(repos);
+            // await processAuthorizePaymentCard({
+            //     project: { typeOf: transaction.project.typeOf, id: transaction.project.id },
+            //     agent: { id: transaction.agent.id },
+            //     object: {
+            //         amount: transaction.object.amount,
+            //         typeOf: factory.action.authorize.paymentMethod.any.ResultType.Payment,
+            //         // typeOf: fromLocation.typeOf,
+            //         paymentMethod: fromLocation.typeOf,
+            //         fromLocation: fromLocation,
+            //         ...{
+            //             description: transaction.object.description
+            //         }
+            //     },
+            //     purpose: { typeOf: transaction.typeOf, id: transaction.id }
+            // })(repos);
         }
     };
 }
@@ -200,9 +212,9 @@ function fixToLocation(
         let toLocation: factory.transaction.moneyTransfer.IToLocation = params.object.toLocation;
 
         if (typeof toLocation.typeOf === 'string') {
-            toLocation = <any>{
+            toLocation = {
                 typeOf: params.object.toLocation.typeOf,
-                identifier: (<any>params.object.toLocation).identifier
+                identifier: params.object.toLocation.identifier
             };
         } else {
             toLocation = <any>{
@@ -230,12 +242,13 @@ export type IAuthorizeOperation<T> = (repos: {
 function processAuthorizePaymentCard(params: {
     project: factory.project.IProject;
     agent: { id: string };
-    object: factory.action.authorize.paymentMethod.any.IObject & {
-        fromLocation?: factory.action.authorize.paymentMethod.any.IPaymentCard;
+    object: factory.action.authorize.offer.monetaryAmount.IObject & {
+        fromLocation?: factory.action.transfer.moneyTransfer.IPaymentCard;
         currency?: string;
     };
-    purpose: factory.action.authorize.paymentMethod.any.IPurpose;
-}): IAuthorizeOperation<factory.action.authorize.paymentMethod.any.IAction> {
+    purpose: factory.action.authorize.offer.monetaryAmount.IPurpose;
+}): IAuthorizeOperation<factory.action.authorize.offer.monetaryAmount.IAction> {
+    // tslint:disable-next-line:max-func-body-length
     return async (repos: {
         action: ActionRepo;
         project: ProjectRepo;
@@ -259,16 +272,29 @@ function processAuthorizePaymentCard(params: {
         }
 
         // 承認アクションを開始する
-        const actionAttributes: factory.action.authorize.paymentMethod.any.IAttributes = {
+        const actionAttributes: factory.action.authorize.offer.monetaryAmount.IAttributes = {
             project: transaction.project,
             typeOf: factory.actionType.AuthorizeAction,
             object: {
-                ...params.object,
-                ...(params.object.fromLocation !== undefined)
-                    ? { accountId: params.object.fromLocation.identifier }
-                    : {},
-                paymentMethod: <string>params.object?.fromLocation?.typeOf,
-                typeOf: factory.action.authorize.paymentMethod.any.ResultType.Payment
+                project: { typeOf: transaction.project.typeOf, id: transaction.project.id },
+                typeOf: factory.chevre.offerType.Offer,
+                itemOffered: {
+                    typeOf: 'MonetaryAmount',
+                    currency: params.object.itemOffered.currency,
+                    value: params.object.itemOffered.value
+                },
+                seller: {
+                    ...transaction.seller,
+                    name: (typeof transaction.seller.name === 'string')
+                        ? transaction.seller.name
+                        : String(transaction.seller.name?.ja)
+                },
+                price: 0,
+                priceCurrency: factory.priceCurrency.JPY,
+                // typeOf: factory.actionType.MoneyTransfer,
+                // amount: params.object.amount,
+                toLocation: params.object.toLocation
+                // pendingTransaction: responseBody
             },
             agent: transaction.agent,
             recipient: recipient,
@@ -277,15 +303,28 @@ function processAuthorizePaymentCard(params: {
         const action = await repos.action.start(actionAttributes);
 
         // 口座取引開始
-        let pendingTransaction: factory.action.authorize.paymentMethod.any.IPendingTransaction;
+        let responseBody: factory.action.authorize.offer.monetaryAmount.IResponseBody;
 
         try {
-            pendingTransaction = await processMoneyTransferTransaction({
+            responseBody = await processMoneyTransferTransaction({
                 project: project,
                 object: params.object,
                 recipient: recipient,
                 transaction: transaction
             });
+
+            // アクションにchevre取引情報を保管
+            await repos.action.actionModel.findByIdAndUpdate(
+                action.id,
+                {
+                    'object.itemOffered.currency': responseBody.object.amount.currency,
+                    'object.pendingTransaction': {
+                        typeOf: responseBody.typeOf,
+                        id: responseBody.id
+                    }
+                }
+            )
+                .exec();
         } catch (error) {
             try {
                 // tslint:disable-next-line:max-line-length no-single-line-block-comment
@@ -300,46 +339,28 @@ function processAuthorizePaymentCard(params: {
             throw error;
         }
 
-        const actionResult: factory.action.authorize.paymentMethod.any.IResult = {
-            accountId: (params.object.fromLocation !== undefined)
-                ? params.object.fromLocation.identifier
-                : '',
-            amount: params.object.amount,
-            paymentMethod: <string>params.object.fromLocation?.typeOf,
-            paymentStatus: factory.paymentStatusType.PaymentDue,
-            paymentMethodId: pendingTransaction.id,
-            name: (typeof params.object.name === 'string')
-                ? params.object.name
-                : (params.object.fromLocation !== undefined)
-                    ? String(params.object.fromLocation.typeOf)
-                    : '',
-            additionalProperty: (Array.isArray(params.object.additionalProperty)) ? params.object.additionalProperty : [],
-            pendingTransaction: pendingTransaction,
-            totalPaymentDue: {
-                typeOf: 'MonetaryAmount',
-                currency: factory.priceCurrency.JPY,
-                value: params.object.amount
-            },
-            ...(params.object.fromLocation !== undefined) ? { fromLocation: params.object.fromLocation } : {},
-            ...(params.object.toLocation !== undefined) ? { toLocation: params.object.toLocation } : {},
-            typeOf: factory.action.authorize.paymentMethod.any.ResultType.Payment
+        const result: factory.action.authorize.offer.monetaryAmount.IResult = {
+            price: 0,
+            priceCurrency: factory.priceCurrency.JPY,
+            // requestBody: requestBody,
+            responseBody: responseBody
         };
 
-        return repos.action.complete({ typeOf: action.typeOf, id: action.id, result: actionResult });
+        return repos.action.complete({ typeOf: action.typeOf, id: action.id, result });
     };
 }
 
-// tslint:disable-next-line:max-func-body-length
+// tslint:disable-next-line:cyclomatic-complexity max-func-body-length
 async function processMoneyTransferTransaction(params: {
     project: factory.project.IProject;
-    object: factory.action.authorize.paymentMethod.any.IObject & {
-        fromLocation?: factory.action.authorize.paymentMethod.any.IPaymentCard;
+    object: factory.action.authorize.offer.monetaryAmount.IObject & {
+        fromLocation?: factory.action.transfer.moneyTransfer.IPaymentCard;
         currency?: string;
     };
     recipient: factory.transaction.moneyTransfer.IRecipient | factory.transaction.placeOrder.ISeller;
     transaction: factory.transaction.ITransaction<factory.transactionType>;
-}): Promise<factory.action.authorize.paymentMethod.any.IPendingTransaction> {
-    let pendingTransaction: factory.action.authorize.paymentMethod.any.IPendingTransaction;
+}): Promise<factory.action.authorize.offer.monetaryAmount.IResponseBody> {
+    let pendingTransaction: factory.action.authorize.offer.monetaryAmount.IResponseBody;
 
     const transaction = params.transaction;
 
@@ -353,18 +374,15 @@ async function processMoneyTransferTransaction(params: {
     const recipient = {
         typeOf: params.recipient.typeOf,
         id: params.recipient.id,
-        name: (typeof (<any>params.recipient).name === 'string')
-            ? (<any>params.recipient).name
-            : ((<any>params.recipient).name !== undefined
-                && (<any>params.recipient).name !== null
-                && typeof (<any>params.recipient).name.ja === 'string')
-                ? (<any>params.recipient).name.ja
+        name: (typeof params.recipient.name === 'string')
+            ? params.recipient.name
+            : (typeof params.recipient.name?.ja === 'string')
+                ? params.recipient.name.ja
                 : `${transaction.typeOf} Transaction ${transaction.id}`,
         ...(typeof params.recipient.url === 'string') ? { url: params.recipient.url } : undefined
     };
 
-    // const description = (typeof params.object.notes === 'string') ? params.object.notes : `for transaction ${transaction.id}`;
-    const description = `Transaction ${transaction.id}`;
+    const description = (typeof params.object.description === 'string') ? params.object.description : `${transaction.typeOf}:${transaction.id}`;
 
     // 最大1ヵ月のオーソリ
     const expires = moment()
@@ -387,8 +405,8 @@ async function processMoneyTransferTransaction(params: {
             object: {
                 amount: {
                     typeOf: 'MonetaryAmount',
-                    value: params.object.amount,
-                    currency: chevre.factory.priceCurrency.JPY
+                    value: params.object.itemOffered.value,
+                    currency: params.object.itemOffered.currency
                 },
                 description: description,
                 fromLocation: {
@@ -403,7 +421,12 @@ async function processMoneyTransferTransaction(params: {
                     typeOf: factory.pecorino.transactionType.Withdraw,
                     id: '' // 空でok
                 }
-            }
+            },
+            // ユニークネスを保証するために識別子を指定する
+            ...(transaction.typeOf === factory.transactionType.MoneyTransfer
+                && typeof transaction.object.pendingTransaction?.identifier === 'string')
+                ? { identifier: transaction.object.pendingTransaction.identifier }
+                : undefined
         });
     } else if (params.object.fromLocation !== undefined && params.object.toLocation !== undefined) {
         pendingTransaction = await moneyTransferService.start({
@@ -415,8 +438,8 @@ async function processMoneyTransferTransaction(params: {
             object: {
                 amount: {
                     typeOf: 'MonetaryAmount',
-                    value: params.object.amount,
-                    currency: chevre.factory.priceCurrency.JPY
+                    value: params.object.itemOffered.value,
+                    currency: params.object.itemOffered.currency
                 },
                 description: description,
                 fromLocation: {
@@ -431,7 +454,12 @@ async function processMoneyTransferTransaction(params: {
                     typeOf: factory.pecorino.transactionType.Transfer,
                     id: '' // 空でok
                 }
-            }
+            },
+            // ユニークネスを保証するために識別子を指定する
+            ...(transaction.typeOf === factory.transactionType.MoneyTransfer
+                && typeof transaction.object.pendingTransaction?.identifier === 'string')
+                ? { identifier: transaction.object.pendingTransaction.identifier }
+                : undefined
         });
     } else if (params.object.fromLocation === undefined && params.object.toLocation !== undefined) {
         pendingTransaction = await moneyTransferService.start({
@@ -443,8 +471,8 @@ async function processMoneyTransferTransaction(params: {
             object: {
                 amount: {
                     typeOf: 'MonetaryAmount',
-                    value: params.object.amount,
-                    currency: chevre.factory.priceCurrency.JPY
+                    value: params.object.itemOffered.value,
+                    currency: params.object.itemOffered.currency
                 },
                 description: description,
                 fromLocation: {
@@ -459,7 +487,12 @@ async function processMoneyTransferTransaction(params: {
                     typeOf: factory.pecorino.transactionType.Deposit,
                     id: '' // 空でok
                 }
-            }
+            },
+            // ユニークネスを保証するために識別子を指定する
+            ...(transaction.typeOf === factory.transactionType.MoneyTransfer
+                && typeof transaction.object.pendingTransaction?.identifier === 'string')
+                ? { identifier: transaction.object.pendingTransaction.identifier }
+                : undefined
         });
     } else {
         throw new factory.errors.Argument('Object', 'At least one of accounts from and to must be specified');
@@ -604,9 +637,9 @@ export function exportTasksById(params: {
 
             case factory.transactionStatusType.Canceled:
             case factory.transactionStatusType.Expired:
-                const voidPaymentTaskAttributes: factory.task.IAttributes<factory.taskName.VoidPayment> = {
+                const voidMoneyTransferTaskAttributes: factory.task.IAttributes<factory.taskName.VoidMoneyTransfer> = {
                     project: { typeOf: transaction.project.typeOf, id: transaction.project.id },
-                    name: factory.taskName.VoidPayment,
+                    name: factory.taskName.VoidMoneyTransfer,
                     status: factory.taskStatus.Ready,
                     runsAt: taskRunsAt,
                     remainingNumberOfTries: 10,
@@ -618,7 +651,7 @@ export function exportTasksById(params: {
                     }
                 };
 
-                taskAttributes.push(voidPaymentTaskAttributes);
+                taskAttributes.push(voidMoneyTransferTaskAttributes);
 
                 break;