+++ CHANGELOG.md
@@ -18,6 +18,22 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Security
 
+## v10.15.0 - 2021-01-26
+
+### Added
+
+- プロジェクトサービスを追加
+
+### Changed
+
+- 通貨転送取引を再実装
+- 注文取引に対する特典口座番号を発行できるように調整
+- 注文作成後の通知に注文トークンを付加
+- Chevre決済中止処理のエラーハンドリングを調整
+- Chevre予約の追加特性からpaymentSeatIndex,csvCode,transactionを削除
+- update @chevre/api-nodejs-client
+- update @cinerino/factory
+
 ## v10.14.1 - 2020-12-17
 
 ### Changed